many cleanups

+++ setup-neovim.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
+
+if [ -n "${SKIP_SETUP_SCRIPTS}" ]; then
+  exit 0
+fi
 rm -rf $HOME/.local/share/lunarvim.old/
 LV_BRANCH='release-1.3/neovim-0.9' bash <(curl -s https://raw.githubusercontent.com/lunarvim/lunarvim/master/utils/installer/install.sh)
